+++ main/include/sensor-task.h
@@ -27,7 +27,7 @@
 #define I2C_MASTER_FREQ_HZ  100000
 
 
-uint16_t get_temp_press(void);
+uint8_t get_temp_press(void);
 
 void init_sensor();
